Move AsyncRock class to it's own file
